+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   request now results in a `404` response, rather than logging an authenticator
   not enabled message. CNJR-5854
 
+### Changed
+- Do not increase secret's value version in case there is no actual change in
+  the secret's value. CNJR-7680
+
 ## [1.21.3] - 2024-12-16
 
 ### Fixed